refactor(archtest): merge 8 saga_*_test.go into saga_invariants_test.go + add consolidation guard (#1213)
Consolidates the 8 saga-theme archtest files into a single
tools/archtest/saga_invariants_test.go per .claude/rules/gocell/ai-robust.md
§"archtest 文件命名" (同主题规则 ≥ 3 → {theme}_invariants_test.go). Pure
relocation: every test function name and per-rule godoc preserved verbatim
(go test -list diff: 0 lost/renamed, +3 new guard tests).

Also upgrades the file-naming convention from Soft to a machine-enforced
Medium for the saga theme: SAGA-INVARIANTS-FILE-CONSOLIDATED-01 (self-hosted
in the merged file) fails if any tools/archtest/*_test.go other than
saga_invariants_test.go declares a SAGA-* invariant — closing the recurrence
vector that let the saga theme fragment into 8 files. AI-robust grade Medium
(Hard infeasible: file-grouping is a meta-property over a file set). Repo-wide
generalization to every theme tracked in #1279.

Closes #1213
Refs #1279
